backend/requirements.txt
flask
flask_cors
dotenv
google-generativeai
